Fix the 'ghdocs' command of utils/psrelease
By default, 'rsync' decides whether to transfer a file by checking its
modification time and size. This can fail to detect differences when
the source and destination files are the same size, as can happen in
'psrelease ghdocs'. The '-c' option of 'rsync' compares files by
checksum, which is more reliable.